requirements.txt
uvloop==0.19.0
fastapi==0.111.0
uvicorn==0.29.0
PyJWT==2.8.0
python-dotenv==1.0.1
pydantic==2.7.1